+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 ~~~~~~
 
 1) [TASK][BREAKING] Modify default deploy task.
+2) [TASK] Add "deploy-fast" task without database and buffer protections - good to deploy to staging instances.
 
 6.2.1
 ~~~~~